Unify sendBotResponse
