+++ packages/server/src/cds/routes.ts
@@ -1,11 +1,11 @@
 // SPDX-FileCopyrightText: Copyright Orangebot, Inc. and Medplum contributors
 // SPDX-License-Identifier: Apache-2.0
-import { allOk, badRequest, ContentType, getStatus, Operator } from '@medplum/core';
+import { ContentType, Operator } from '@medplum/core';
 import type { Bot } from '@medplum/fhirtypes';
 import type { Request, Response } from 'express';
 import { Router } from 'express';
 import { executeBot } from '../bots/execute';
-import { getBotDefaultHeaders, getResponseBodyFromResult } from '../bots/utils';
+import { getBotDefaultHeaders, sendBotResponse } from '../bots/utils';
 import { getAuthenticatedContext } from '../context';
 import { authenticateRequest } from '../oauth/middleware';
 
@@ -75,7 +75,5 @@ cdsRouter.post('/:id', async (req: Request, res: Response) => {
     defaultHeaders: getBotDefaultHeaders(req, bot),
   });
 
-  const responseBody = getResponseBodyFromResult(result);
-  const outcome = result.success ? allOk : badRequest(result.logResult);
-  res.status(getStatus(outcome)).type(ContentType.JSON).send(responseBody);
+  await sendBotResponse(req, res, result);
 });
